+++ src/error.rs
@@ -7,38 +7,41 @@ use failure::{Context, Fail};
 #[derive(Debug, Fail)]
 pub enum CeleryError {
     /// The queue you're attempting to use has not been defined.
-    #[fail(display = "Unknown queue '{}'", _0)]
+    #[fail(display = "CeleryError: unknown queue '{}'", _0)]
     UnknownQueue(String),
 
     /// Raised when `Celery::consume_from` is given an empty array of queues.
-    #[fail(display = "At least one queue required to consume from")]
+    #[fail(display = "CeleryError: at least one queue required to consume from")]
     NoQueueToConsume,
 
     /// Forced shutdown.
-    #[fail(display = "Forced shutdown")]
+    #[fail(display = "CeleryError: forced shutdown")]
     ForcedShutdown,
 
     /// Any other broker-level error that could happen when initializing.
-    #[fail(display = "{}", _0)]
+    #[fail(display = "CeleryError: broker error ({})", _0)]
     BrokerError(#[fail(cause)] BrokerError),
 
     /// Any other IO error that could occur.
-    #[fail(display = "{}", _0)]
+    #[fail(display = "CeleryError: IO error ({})", _0)]
     IoError(#[fail(cause)] std::io::Error),
 
     /// A protocol error.
-    #[fail(display = "Protocol error: {}", _0)]
+    #[fail(display = "CeleryError: protocol error ({})", _0)]
     ProtocolError(#[fail(cause)] ProtocolError),
 
     /// An invalid glob pattern for a routing rule.
-    #[fail(display = "Invalid glob routing rule '{}'", _0)]
+    #[fail(display = "CeleryError: invalid glob routing rule '{}'", _0)]
     BadRoutingPattern(#[fail(cause)] globset::Error),
 
     /// There is already a task registerd to this name.
-    #[fail(display = "There is already a task registered as '{}'", _0)]
+    #[fail(
+        display = "CeleryError: there is already a task registered as '{}'",
+        _0
+    )]
     TaskRegistrationError(String),
 
-    #[fail(display = "Received unregistered task {}", _0)]
+    #[fail(display = "CeleryError: received unregistered task {}", _0)]
     UnregisteredTaskError(String),
 }
 
@@ -55,7 +58,7 @@ pub enum TaskError {
     /// If that service is temporarily unavailable the task should raise an `ExpectedError`.
     ///
     /// Tasks are always retried with capped exponential backoff.
-    #[fail(display = "{}", _0)]
+    #[fail(display = "TaskError (expected): {}", _0)]
     ExpectedError(String),
 
     /// Should be used when a task encounters an error that is unexpected.
@@ -64,7 +67,7 @@ pub enum TaskError {
     /// when this error is encountered is determined by the
     /// [`TaskOptions::retry_for_unexpected`](../task/struct.TaskOptions.html#structfield.retry_for_unexpected)
     /// setting.
-    #[fail(display = "{}", _0)]
+    #[fail(display = "TaskError: {}", _0)]
     UnexpectedError(String),
 
     /// Raised when a task runs over its time limit specified by the
@@ -76,7 +79,7 @@ pub enum TaskError {
     /// Typically a task implementation doesn't need to return these errors directly
     /// because they will be raised automatically when the task runs over it's `timeout`,
     /// provided the task yields control at some point (like with non-blocking IO).
-    #[fail(display = "Task timed out")]
+    #[fail(display = "TaskError: task timed out")]
     TimeoutError,
 }
 
@@ -84,15 +87,15 @@ pub enum TaskError {
 #[derive(Debug, Fail)]
 pub(crate) enum TraceError {
     /// Raised when a task throws an error while executing.
-    #[fail(display = "Task failed with {}", _0)]
+    #[fail(display = "TraceError: task failed with {}", _0)]
     TaskError(TaskError),
 
     /// Raised when an expired task is received.
-    #[fail(display = "Task expired")]
+    #[fail(display = "TraceError: task expired")]
     ExpirationError,
 
     /// Raised when a task should be retried.
-    #[fail(display = "Retrying task")]
+    #[fail(display = "retrying task")]
     Retry(Option<DateTime<Utc>>),
 }
 
@@ -100,23 +103,23 @@ pub(crate) enum TraceError {
 #[derive(Debug, Fail)]
 pub enum BrokerError {
     /// Raised when a broker URL can't be parsed.
-    #[fail(display = "Invalid broker URL: {}", _0)]
+    #[fail(display = "BrokerError: invalid broker URL '{}'", _0)]
     InvalidBrokerUrl(String),
 
     /// The queue you're attempting to use has not been defined.
-    #[fail(display = "Unknown queue '{}'", _0)]
+    #[fail(display = "BrokerError: unknown queue '{}'", _0)]
     UnknownQueue(String),
 
     /// Broker is disconnected.
-    #[fail(display = "Broker not connected")]
+    #[fail(display = "BrokerError: broker not connected")]
     NotConnected,
 
     /// Any IO error that could occur.
-    #[fail(display = "{}", _0)]
+    #[fail(display = "BrokerError: IO error ({})", _0)]
     IoError(#[fail(cause)] std::io::Error),
 
     /// Any other AMQP error that could happen.
-    #[fail(display = "{}", _0)]
+    #[fail(display = "BrokerError: AMQP error ({})", _0)]
     AMQPError(#[fail(cause)] lapin::Error),
 }
 
@@ -124,19 +127,19 @@ pub enum BrokerError {
 #[derive(Debug, Fail)]
 pub enum ProtocolError {
     /// Raised when a required message property is missing.
-    #[fail(display = "Missing required property '{}'", _0)]
+    #[fail(display = "ProtocolError: missing required property '{}'", _0)]
     MissingRequiredProperty(String),
 
     /// Raised when the headers are missing altogether.
-    #[fail(display = "Missing headers")]
+    #[fail(display = "ProtocolError: missing headers")]
     MissingHeaders,
 
     /// Raised when a required message header is missing.
-    #[fail(display = "Missing required property '{}'", _0)]
+    #[fail(display = "ProtocolError: missing required property '{}'", _0)]
     MissingRequiredHeader(String),
 
     /// Raised when serializing or de-serializing a message body fails.
-    #[fail(display = "{}", _0)]
+    #[fail(display = "ProtocolError: serialization error ({})", _0)]
     BodySerializationError(#[fail(cause)] serde_json::Error),
 }
 
@@ -173,9 +176,10 @@ impl From<serde_json::Error> for ProtocolError {
 impl From<lapin::Error> for BrokerError {
     fn from(err: lapin::Error) -> Self {
         match err {
-            lapin::Error::IOError(e) => {
-                BrokerError::IoError(std::io::Error::new((*e).kind(), format!("{}", *e)))
-            }
+            lapin::Error::IOError(e) => BrokerError::IoError(std::io::Error::new(
+                (*e).kind(),
+                format!("{} from AMQP broker", *e),
+            )),
             _ => BrokerError::AMQPError(err),
         }
     }