+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `BrokerError` variants trimmed and simplified.
 - The error handler closure passed to `Broker::consume` now takes a `BrokerError` as an argument.
+- Improved error messages.
 
 ## v0.2.6 - 2019-05-14